+++ C_Sharp/Client/WoM_BlazorApp/Services/Interfaces/ISaleService.cs
@@ -1,4 +1,4 @@
-using ApiContracts.Sales;
+using ApiContracts;
 
 namespace WoM_BlazorApp.Http;
 
@@ -8,5 +8,5 @@ public interface ISaleService
     Task<SaleDto> GetByIdAsync(int id);
     Task<SaleDto> CreateAsync(CreateSaleDto dto);
     Task UpdateAsync(int id, UpdateSaleDto dto);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
